changed type of amount in entries and transfers to int

+++ db/migration/000001_init_schema.up.sql
@@ -10,7 +10,7 @@ create table accounts(
 create table entries(
     id bigserial primary key,
     account_id bigint references accounts(id) on delete cascade not null,
-    amount bigint not null,
+    amount int not null,
     created_at timestamptz default now() not null,
     updated_at timestamptz default now() not null,
     deleted_at timestamptz default '0001-01-01 00:00:00Z' not null
@@ -22,7 +22,7 @@ create table transfers(
     to_account_id bigint references accounts(id) on delete cascade not null,
     incoming_entry_id bigint references entries(id) on delete cascade not null,
     outgoing_entry_id bigint references entries(id) on delete cascade not null,
-    amount bigint not null,
+    amount int not null,
     created_at timestamptz default now() not null,
     updated_at timestamptz default now() not null,
     deleted_at timestamptz default '0001-01-01 00:00:00Z' not null
